NEW - # 563: drag and drop of text jumps to bottom of editor http://trac.cfeclipse.org/ticket/563

+++ svn/org.cfeclipse.cfml/trunk/src/org/cfeclipse/cfml/editors/CFMLEditor.java
@@ -1268,12 +1268,14 @@ IReconcilingParticipant, IProjectionListener, IPropertyChangeListener, IShowInSo
 			return;
 
 		StyledText st= getSourceViewer().getTextWidget();
-		dndService.removeMergedDropTarget(st);
-
-		DragSource dragSource= (DragSource)st.getData(DND.DRAG_SOURCE_KEY);
-		if (dragSource != null) {
-			dragSource.dispose();
-			st.setData(DND.DRAG_SOURCE_KEY, null);
+		if(st != null) {			
+			dndService.removeMergedDropTarget(st);
+			
+			DragSource dragSource= (DragSource)st.getData(DND.DRAG_SOURCE_KEY);
+			if (dragSource != null) {
+				dragSource.dispose();
+				st.setData(DND.DRAG_SOURCE_KEY, null);
+			}
 		}
 		//Remove the listener
 		if(SelectionCursorListener!=null){
